update local repo to newest

+++ src/tools/rbd/action/Remove.cc
@@ -5,12 +5,33 @@
 #include "tools/rbd/Shell.h"
 #include "tools/rbd/Utils.h"
 #include "common/errno.h"
+#include "include/stringify.h"
 #include <iostream>
 #include <boost/program_options.hpp>
 
 namespace rbd {
 namespace action {
 namespace remove {
+
+namespace {
+
+bool is_auto_delete_snapshot(librbd::Image* image,
+                             const librbd::snap_info_t &snap_info) {
+  librbd::snap_namespace_type_t namespace_type;
+  int r = image->snap_get_namespace_type(snap_info.id, &namespace_type);
+  if (r < 0) {
+    return false;
+  }
+
+  switch (namespace_type) {
+  case RBD_SNAP_NAMESPACE_TYPE_TRASH:
+    return true;
+  default:
+    return false;
+  }
+}
+
+} // anonymous namespace
 
 namespace at = argument_types;
 namespace po = boost::program_options;
@@ -34,7 +55,8 @@ void get_arguments(po::options_description *positional,
   at::add_no_progress_option(options);
 }
 
-int execute(const po::variables_map &vm) {
+int execute(const po::variables_map &vm,
+            const std::vector<std::string> &ceph_global_init_args) {
   size_t arg_index = 0;
   std::string pool_name;
   std::string image_name;
@@ -60,9 +82,30 @@ int execute(const po::variables_map &vm) {
                 vm[at::NO_PROGRESS].as<bool>());
   if (r < 0) {
     if (r == -ENOTEMPTY) {
-      std::cerr << "rbd: image has snapshots - these must be deleted"
-                << " with 'rbd snap purge' before the image can be removed."
-                << std::endl;
+      librbd::Image image;
+      std::vector<librbd::snap_info_t> snaps;
+      r = utils::open_image(io_ctx, image_name, true, &image);
+      if (r >= 0) {
+        r = image.snap_list(snaps);
+      }
+      if (r >= 0) {
+        snaps.erase(std::remove_if(snaps.begin(), snaps.end(),
+			           [&image](const librbd::snap_info_t& snap) {
+                                     return is_auto_delete_snapshot(&image,
+                                                                    snap);
+                                   }),
+                    snaps.end());
+      }
+
+      if (!snaps.empty()) {
+        std::cerr << "rbd: image has snapshots - these must be deleted"
+                  << " with 'rbd snap purge' before the image can be removed."
+                  << std::endl;
+      } else {
+        std::cerr << "rbd: image has snapshots with linked clones - these must "
+                  << "be deleted or flattened before the image can be removed."
+                  << std::endl;
+      }
     } else if (r == -EBUSY) {
       std::cerr << "rbd: error: image still has watchers"
                 << std::endl
@@ -73,18 +116,27 @@ int execute(const po::variables_map &vm) {
     } else if (r == -EMLINK) {
       librbd::Image image;
       int image_r = utils::open_image(io_ctx, image_name, true, &image);
-      librbd::group_spec_t group_spec;
+      librbd::group_info_t group_info;
       if (image_r == 0) {
-	image_r = image.get_group(&group_spec);
+	image_r = image.get_group(&group_info, sizeof(group_info));
       }
-      if (image_r == 0)
-	std::cerr << "rbd: error: image belongs to a consistency group "
-		  << group_spec.pool << "." << group_spec.name;
-      else
-	std::cerr << "rbd: error: image belongs to a consistency group";
+      if (image_r == 0) {
+        std::string pool_name = "";
+        librados::Rados rados(io_ctx);
+        librados::IoCtx pool_io_ctx;
+        r = rados.ioctx_create2(group_info.pool, pool_io_ctx);
+        if (r < 0) {
+          pool_name = "<missing data pool " + stringify(group_info.pool) + ">";
+        } else {
+          pool_name = pool_io_ctx.get_pool_name();
+        }
+        std::cerr << "rbd: error: image belongs to a group "
+                  << pool_name << "/" << group_info.name;
+      } else
+	std::cerr << "rbd: error: image belongs to a group";
 
       std::cerr << std::endl
-		<< "Remove the image from the consistency group and try again."
+		<< "Remove the image from the group and try again."
 		<< std::endl;
       image.close();
     } else {
